modularize subagent runtime by responsibility
